+++ srm-mutuelle-backend/src/main/java/ma/srm/mutuelle/service/ContractedDoctorService.java
@@ -61,7 +61,7 @@ public class ContractedDoctorService {
 
 	@Transactional
 	public void delete(Long id, AppUser user) {
-		AccessRules.assertStaffWrite(user);
+		AccessRules.assertAdmin(user);
 		contractedDoctorRepository.deleteById(id);
 	}
