+++ srm-mutuelle-backend/src/main/java/ma/srm/mutuelle/service/SpecialDiseaseService.java
@@ -73,7 +73,7 @@ public class SpecialDiseaseService {
 
 	@Transactional
 	public void delete(Long id, AppUser user) {
-		AccessRules.assertStaffWrite(user);
+		AccessRules.assertAdmin(user);
 		repository.deleteById(id);
 	}
